# api-node-express/openapi.yaml
openapi: 3.0.3
info:
  title: ApiBattle API
  version: 1.0.0
  description: |
    Banking ledger API for managing accounts and transactions.

    Accounts are provisioned in the database (seeded accounts: 1000 and 2000).

    Transactions are append-only: they can be added but never modified or
    deleted. An account's balance is always derived from its transactions
    (credits minus debits). Amounts are integers in minor units (e.g. cents).
servers:
  - url: /
security: []  # The API has no authentication.
tags:
  - name: Accounts
    description: Account management and balances
  - name: Transactions
    description: Append-only ledger entries
  - name: Health
    description: Service health

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: health
      responses:
        "200":
          description: Service is up
          content:
            text/plain:
              schema:
                type: string
                example: ok

  /accounts:
    get:
      tags: [Accounts]
      summary: List accounts with their current balances (ordered by ID)
      operationId: listAccounts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /accounts/{id}:
    get:
      tags: [Accounts]
      summary: Get account details including current balance
      operationId: getAccount
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  /accounts/{id}/transactions:
    get:
      tags: [Transactions]
      summary: List account transactions (latest first)
      operationId: listTransactions
      parameters:
        - $ref: "#/components/parameters/AccountId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Transactions]
      summary: Add a transaction to an account
      description: Debits that would make the balance negative are rejected with 422.
      operationId: addTransaction
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: insufficient funds
        "500":
          $ref: "#/components/responses/InternalError"

  /transactions/{id}:
    get:
      tags: [Transactions]
      summary: Get transaction details
      operationId: getTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  parameters:
    AccountId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:
    Account:
      type: object
      required: [id, name, balance, created_at]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Alice
        balance:
          type: integer
          format: int64
          description: Credits minus debits, in minor units
          example: 7500
        created_at:
          type: string
          format: date-time
          example: "2026-10-03T02:44:35.274Z"

    AccountPage:
      type: object
      required: [data, page, page_size, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Account"
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          description: Total number of accounts
          example: 2

    CreateTransactionRequest:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: integer
          format: int64
          minimum: 1
          description: Amount in minor units (e.g. cents)
          example: 10000
        description:
          type: string
          maxLength: 255
          nullable: true
          example: Salary

    Transaction:
      type: object
      required: [id, account_id, type, amount, description, created_at]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        account_id:
          type: integer
          format: int64
          example: 1
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: integer
          format: int64
          example: 10000
        description:
          type: string
          nullable: true
          example: Salary
        created_at:
          type: string
          format: date-time
          example: "2026-10-03T02:44:35.294Z"

    TransactionPage:
      type: object
      required: [data, page, page_size, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          description: Total number of transactions for the account
          example: 2

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "'type' must be either 'credit' or 'debit'"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: account not found
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: internal server error